details::[storages/aheui_states] 템플릿 구조체 추가
전자는 아희에서 제공하는 28개의 저장공간을 묶으며, 후자는 현재 아희 코드의 상태를 저장합니다.

+++ include/TemplatedAH/string.hpp
@@ -135,7 +135,7 @@ namespace tah
 		template<char32_t Jongsung_>
 		struct get_jongsung_index<Jongsung_, sizeof(hangul::jongsungs) / sizeof(char32_t)>
 		{
-			static constexpr std::size_t value = -1;
+			static constexpr std::size_t value = static_cast<std::size_t>(-1);
 		};
 	}
 }
